fix: cross-platform routing, TLS, and 32-bit correctness pass
CI/build:

- Fix Zig 0.16 Linux/macOS compile errors: replace removed std.c.stat with statx (Linux) plus versioned custom extern stat (macOS/other), matching the existing static-file serving approach; fix mtime field names (sec/nsec).

- Fix variadic fcntl call (cast 0 to c_int) for Zig 0.16.

- Make all atomic counters 32-bit safe (u64 atomics -> usize atomics, u64 public API preserved via casts) so x86 32-bit compiles; verified x86-linux-gnu. Fixes pool, worker pool, DNS cache, metrics Counter/Gauge, watcher, multipart boundary counter, unknownCount, liveReloadEventId.

- Fix QUIC stream 32-bit slice-index type (u64 offset -> usize cast).

Router/canonical routing:

- Request path sanitization: strip query/fragment for matching, populate Context.query, preserve peer/TLS/trust fields across match, reject query/fragment in route registration (InvalidPattern).

- Fix queryParam (was always empty on real server requests because transports strip queries): now reads Context.query first with path fallback.

- HEAD falls back to GET handler (explicit HEAD still wins); original method preserved.

- hasConflict/remove ignore query strings; duplicate detection stays shape-based (nested slugs, different param names still conflict).

- Wildcard remainder excludes query and preserves nested slashes.

Docs/OpenAPI/GraphQL reserved routes:

- Add Metadata.internal flag; docs UI mounts all pages/assets as internal; OpenAPI generator skips internal routes and only filters exact/slash-boundary docs prefixes (user routes like /docs-custom, /redocly no longer hidden); custom docs mounts excluded regardless of path.

- Docs mount rolls back partial registration on error; GraphQL mount pre-checks POST/GET/OPTIONS atomically with rollback.

- Disabled features reserve nothing: docs mount returns early when disabled, only enabled routes are conflict-checked, so user can reclaim /docs etc.

TLS:

- Client now sends SNI for DNS hosts in all verify modes (previously verify=none sent no SNI, causing TlsAlert on virtual-hosted HTTPS). IP literals still skip SNI per RFC 6066. Chain verification still skipped for verify=none, so sandbox clock jumps stay bypassed.

Verification:

- 467/467 unit tests pass (8 new router/OpenAPI regression tests). Cross-compiles: x86_64-linux-gnu, x86-linux-gnu (32-bit), aarch64-macos ok; x86-windows-gnu lib compiles (test-link of Windows system libs is a sandbox toolchain limit).

- All 82 examples build; all 82 execute (servers timeout-kill after serving, clients exit 0; httpbun.com interop, website embedded routes, GraphQL, SSE/WebSocket/session/metrics/OpenAPI verified).

+++ src/concurrency/worker_pool.zig
@@ -24,11 +24,11 @@ pub const Config = struct {
 };
 
 pub const Stats = struct {
-    submitted: std.atomic.Value(u64) = .init(0),
-    completed: std.atomic.Value(u64) = .init(0),
-    failed: std.atomic.Value(u64) = .init(0),
-    cancelled: std.atomic.Value(u64) = .init(0),
-    rejected: std.atomic.Value(u64) = .init(0),
+    submitted: std.atomic.Value(usize) = .init(0),
+    completed: std.atomic.Value(usize) = .init(0),
+    failed: std.atomic.Value(usize) = .init(0),
+    cancelled: std.atomic.Value(usize) = .init(0),
+    rejected: std.atomic.Value(usize) = .init(0),
 
     pub fn snapshot(self: *const Stats) Stats {
         return .{
